add _calculate_mid_points to utils.py
