+++ dask_geopandas/geohash.py
@@ -1,5 +1,9 @@
 import numpy as np
 import pandas as pd
+from .utils import _calculate_mid_points
+
+
+# Implementation based on deprecated https://pypi.org/project/neathgeohash/#description
 
 
 def _geohash(gdf, precision):
@@ -29,29 +33,6 @@ def _geohash(gdf, precision):
     geohash = encode_geohash(coords, precision)
 
     return pd.Series(geohash, index=gdf.index, name="geohash")
-
-
-def _calculate_mid_points(bounds):
-    """
-    Calculate geohash based on the mid points of each geometry
-
-    Parameters
-    ----------
-    gdf : GeoDataFrame
-    precision : int
-        precision of the Geohash
-
-    Returns
-    ---------
-    x_mids : mid points of x values
-    y_mids : mid points of y values
-    """
-
-    # Calculate mid points for x and y bound coords
-    x_mids = (bounds[:, 0] + bounds[:, 2]) / 2.0
-    y_mids = (bounds[:, 1] + bounds[:, 3]) / 2.0
-
-    return x_mids, y_mids
 
 
 def encode_geohash(coords, precision):
